migration-backend: Remove unused new book nft methods
