+++ migration-backend/pkg/likenft/evm/new_book_nft_with_royalty_and_salt.go
@@ -2,9 +2,12 @@ package evm
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"log/slog"
 	"math/big"
 
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/likecoin/like-migration-backend/pkg/ethereum"
 	"github.com/likecoin/like-migration-backend/pkg/likenft/evm/like_protocol"
@@ -61,4 +64,21 @@ func (l *LikeProtocol) NewBookNFTWithRoyaltyAndSalt(
 	}
 
 	return tx, txReceipt, err
+}
+
+func (l *LikeProtocol) GetClassIdFromNewClassTransaction(txReceipt *types.Receipt) (*common.Address, error) {
+	filterer, err := like_protocol.NewLikeProtocolFilterer(l.ContractAddress, l.Client)
+	if err != nil {
+		return nil, fmt.Errorf("err likenft.NewLikeProtocolFilterer: %v", err)
+	}
+
+	logs := txReceipt.Logs
+
+	for _, vLog := range logs {
+		newClassEvent, err := filterer.ParseNewBookNFT(*vLog)
+		if err == nil {
+			return &newClassEvent.BookNFT, nil
+		}
+	}
+	return nil, errors.New("err finding new book nft event from tx log")
 }
